Convert portal feature flags store to TS
