+++ packages/builder/src/stores/portal/featureFlags.ts
@@ -1,0 +1,19 @@
+import { derived, Readable } from "svelte/store"
+import { auth } from "@/stores/portal"
+
+export const INITIAL_FEATUREFLAG_STATE = {
+  SQS: false,
+  DEFAULT_VALUES: false,
+  BUDIBASE_AI: false,
+  AI_CUSTOM_CONFIGS: false,
+}
+
+export const featureFlags: Readable<Record<string, any>> = derived(
+  [auth],
+  ([$auth]) => {
+    return {
+      ...INITIAL_FEATUREFLAG_STATE,
+      ...($auth?.user?.flags || {}),
+    }
+  }
+)
