+++ packages/types/src/api/web/global/self.ts
@@ -8,5 +8,5 @@ export interface GenerateAPIKeyResponse extends DevInfo {}
 export interface FetchAPIKeyResponse extends DevInfo {}
 
 export interface GetGlobalSelfResponse extends User {
-  flags?: Record<string, string>
+  flags?: Record<string, any>
 }
